+++ Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 let package = Package(
     name: "BatteryTruth",
     platforms: [
-        .macOS(.v15)
+        .macOS(.v14)
     ],
     products: [
         .library(